# pay-guard-backend/docs/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/businesses/{businessId}/transactions/{transactionId}/history": {
      "get": {
        "operationId": "Transactions_history",
        "parameters": [
          {
            "name": "businessId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "transactionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Transactions"
        ]
      }
    },
...
    "/api/v1/businesses/{businessId}/receipt-review-queue": {
      "get": {
        "operationId": "ReceiptReviewCase_list",
        "parameters": [
          {
            "name": "businessId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "OPEN",
                "ACKNOWLEDGED",
                "RESOLVED"
              ]
            }
          },
          {
            "name": "reasonCode",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "NO_QR",
                "MULTIPLE_QR",
                "UNSUPPORTED_PROOF",
                "INCOMPLETE_QR",
                "UNSUPPORTED_BANK",
                "BANK_MISMATCH",
                "REFERENCE_MISMATCH",
                "AMOUNT_MISMATCH",
                "DATE_MISMATCH",
                "TIME_MISMATCH",
                "ACCOUNT_MISMATCH"
              ]
            }
          },
          {
            "name": "branchId",
            "required": false,
            "in": "query",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 50,
              "allOf": [
                {
                  "$ref": "#/components/schemas/Object"
                }
              ]
            }
          },
          {
            "name": "offset",
            "required": true,
            "in": "query",
            "schema": {
              "minimum": 0,
              "maximum": 10000,
              "default": 0,
              "allOf": [
                {
                  "$ref": "#/components/schemas/Object"
                }
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Receipt Review Queue"
        ]
      }
    },
    "/api/v1/businesses/{businessId}/receipt-review-queue/ageing-summary": {
      "get": {
        "operationId": "ReceiptReviewCase_ageingSummary",
        "parameters": [
          {
            "name": "businessId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "reasonCode",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "NO_QR",
                "MULTIPLE_QR",
                "UNSUPPORTED_PROOF",
                "INCOMPLETE_QR",
                "UNSUPPORTED_BANK",
                "BANK_MISMATCH",
                "REFERENCE_MISMATCH",
                "AMOUNT_MISMATCH",
                "DATE_MISMATCH",
                "TIME_MISMATCH",
                "ACCOUNT_MISMATCH"
              ]
            }
          },
          {
            "name": "branchId",
            "required": false,
            "in": "query",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "slaHours",
            "required": true,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 720,
              "default": 24,
              "allOf": [
                {
                  "$ref": "#/components/schemas/Object"
                }
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Receipt Review Queue"
        ]
      }
    },
    "/api/v1/businesses/{businessId}/receipt-review-queue/{caseId}/history": {
      "get": {
        "operationId": "ReceiptReviewCase_history",
        "parameters": [
          {
            "name": "businessId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "caseId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Receipt Review Queue"
        ]
      }
    },
    "/api/v1/businesses/{businessId}/receipt-review-queue/{caseId}/acknowledge": {
      "post": {
        "operationId": "ReceiptReviewCase_acknowledge",
        "parameters": [
          {
            "name": "businessId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "caseId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AcknowledgeReceiptReviewCaseDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Receipt Review Queue"
        ]
      }
    },
    "/api/v1/businesses/{businessId}/receipt-review-queue/{caseId}/resolve": {
      "post": {
        "operationId": "ReceiptReviewCase_resolve",
        "parameters": [
          {
            "name": "businessId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "caseId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResolveReceiptReviewCaseDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Receipt Review Queue"
        ]
      }
    },
...
    "/api/v1/subscription-plans": {
      "get": {
        "operationId": "SubscriptionPurchase_plans",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Subscriptions"
        ]
      }
    },
    "/api/v1/businesses/{businessId}/branches/{branchId}/subscription-purchases": {
      "post": {
        "operationId": "SubscriptionPurchase_create",
        "parameters": [
          {
            "name": "businessId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "branchId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSubscriptionPurchaseDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Subscriptions"
        ]
      },
      "get": {
        "operationId": "SubscriptionPurchase_list",
        "parameters": [
          {
            "name": "businessId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "branchId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "ORDER_CREATED",
                "PROOF_RECEIVED",
                "VERIFICATION_PENDING",
                "VERIFIED",
                "FAILED",
                "DUPLICATE",
                "CANCELLED"
              ]
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 50,
              "allOf": [
                {
                  "$ref": "#/components/schemas/Object"
                }
              ]
            }
          },
          {
            "name": "offset",
            "required": true,
            "in": "query",
            "schema": {
              "minimum": 0,
              "maximum": 10000,
              "default": 0,
              "allOf": [
                {
                  "$ref": "#/components/schemas/Object"
                }
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Subscriptions"
        ]
      }
    },
    "/api/v1/businesses/{businessId}/branches/{branchId}/subscription-purchases/{purchaseId}": {
      "get": {
        "operationId": "SubscriptionPurchase_require",
        "parameters": [
          {
            "name": "businessId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "branchId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "purchaseId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Subscriptions"
        ]
      }
    },
    "/api/v1/businesses/{businessId}/branches/{branchId}/subscription-purchases/{purchaseId}/proof": {
      "post": {
        "operationId": "SubscriptionPurchase_proof",
        "parameters": [
          {
            "name": "businessId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "branchId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "purchaseId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Subscriptions"
        ]
      }
    },
    "/api/v1/businesses/{businessId}/branches/{branchId}/subscription-purchases/{purchaseId}/verify": {
      "post": {
        "operationId": "SubscriptionPurchase_verify",
        "parameters": [
          {
            "name": "businessId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "branchId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "purchaseId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Subscriptions"
        ]
      }
    },
...
  },
  "info": {
    "title": "PayGuard API",
    "description": "Canonical API contract for the PayGuard web portal, Android app, workers, and third-party integrations.",
    "version": "2.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "Local development"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Access token returned by POST /api/v1/auth/login."
      }
    },
    "schemas": {
...
      "Object": {
        "type": "object",
        "properties": {}
      },
...
      "AcknowledgeReceiptReviewCaseDto": {
        "type": "object",
        "properties": {
          "note": {
            "type": "string",
            "minLength": 3,
            "maxLength": 500
          }
        },
        "required": [
          "note"
        ]
      },
      "ResolveReceiptReviewCaseDto": {
        "type": "object",
        "properties": {
          "resolutionCode": {
            "type": "string",
            "enum": [
              "EVIDENCE_REPLACED",
              "FALSE_POSITIVE",
              "INVALID_RECEIPT",
              "DUPLICATE_RECEIPT",
              "OTHER"
            ]
          },
          "note": {
            "type": "string",
            "maxLength": 500
          }
        },
        "required": [
          "resolutionCode"
        ]
      },
...
      "CreateSubscriptionPurchaseDto": {
        "type": "object",
        "properties": {
          "idempotencyKey": {
            "type": "string",
            "format": "uuid"
          },
          "planId": {
            "type": "string",
            "format": "uuid"
          },
          "paymentBankId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "idempotencyKey",
          "planId",
          "paymentBankId"
        ]
      },
...
    }
  }
}